+++ src/config/services.php
@@ -37,6 +37,7 @@ return [
 
     'mercadopago' => [
         'token' => env('MERCADOPAGO_ACCESS_TOKEN'),
+        'webhook_secret' => env('MERCADOPAGO_WEBHOOK_SECRET'),
     ],
 
 ];